lifted down the < sign management in the MatrixCell class
